SS4 Upgrade
* Added namespaces
* Updated directory structures
* Changed to vendor module
