+++ src/SitemapPageController.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace Tractorcow\Sitemap2;
+
+use PageController;
+
+/**
+* @author Damian Mooyman
+*
+* @package sitemap2
+*/
+class SitemapPage_Controller extends PageController
+{
+
+}
